+++ src/main/java/BlackjackApplication.java
@@ -1,8 +1,8 @@
 import controller.BlackjackController;
 import java.util.List;
-import model.Card;
-import model.Deck;
-import model.DeckFactory;
+import model.card.Card;
+import model.deck.Deck;
+import model.deck.DeckFactory;
 import model.Judge;
 import view.InputView;
 import view.OutputView;
