+++ src/main/java/controller/BlackjackController.java
@@ -7,15 +7,14 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-import model.Cards;
-import model.DealerCardsFactory;
-import model.Deck;
+import model.cards.Cards;
+import model.cards.DealerCardsFactory;
+import model.deck.Deck;
 import model.GameResult;
 import model.GameResults;
 import model.Judge;
-import model.PlayerCardsFactory;
+import model.cards.PlayerCardsFactory;
 import model.Players;
-import model.UserInput;
 import view.InputView;
 import view.OutputView;
