+++ src/main/java/model/cards/PlayerCardsFactory.java
@@ -1,6 +1,7 @@
-package model;
+package model.cards;
 
 import java.util.ArrayList;
+import model.deck.Deck;
 
 public class PlayerCardsFactory {
