+++ src/test/java/model/CardsTest.java
@@ -5,6 +5,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.ArrayList;
 import java.util.List;
+import model.card.Card;
+import model.card.CardNumber;
+import model.card.CardShape;
+import model.cards.Cards;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
