+++ src/test/java/model/JudgeTest.java
@@ -4,6 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
+import model.card.Card;
+import model.card.CardNumber;
+import model.card.CardShape;
+import model.cards.Cards;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
